test: split runtime checkpoint resume coverage
